+++ rust-src/concordium_base/src/updates.rs
@@ -796,11 +796,17 @@ impl Serial for UpdatePayload {
             }
             UpdatePayload::AddAnonymityRevoker(add_ar) => {
                 12u8.serial(out);
-                add_ar.serial(out)
+                let mut inner = Vec::new();
+                add_ar.serial(&mut inner);
+                (inner.len() as u32).serial(out);
+                out.write_all(&inner).unwrap();
             }
             UpdatePayload::AddIdentityProvider(add_ip) => {
                 13u8.serial(out);
-                add_ip.serial(out)
+                let mut inner = Vec::new();
+                add_ip.serial(&mut inner);
+                (inner.len() as u32).serial(out);
+                out.write_all(&inner).unwrap();
             }
             UpdatePayload::CooldownParametersCPV1(cp) => {
                 14u8.serial(out);
@@ -836,8 +842,14 @@ impl Deserial for UpdatePayload {
             9u8 => Ok(UpdatePayload::BakerStakeThreshold(source.get()?)),
             10u8 => Ok(UpdatePayload::Root(source.get()?)),
             11u8 => Ok(UpdatePayload::Level1(source.get()?)),
-            12u8 => Ok(UpdatePayload::AddAnonymityRevoker(source.get()?)),
-            13u8 => Ok(UpdatePayload::AddIdentityProvider(source.get()?)),
+            12u8 => {
+                let _len: u32 = source.get()?;
+                Ok(UpdatePayload::AddAnonymityRevoker(source.get()?))
+            }
+            13u8 => {
+                let _len: u32 = source.get()?;
+                Ok(UpdatePayload::AddIdentityProvider(source.get()?))
+            }
             14u8 => Ok(UpdatePayload::CooldownParametersCPV1(source.get()?)),
             15u8 => Ok(UpdatePayload::PoolParametersCPV1(source.get()?)),
             16u8 => Ok(UpdatePayload::TimeParametersCPV1(source.get()?)),